fix(Dockerfile): COPY file with chown for jekyll:jekyll
docker build -t beautiful-jekyll "$PWD" break with the following error:

There was an error while trying to write to `/srv/jekyll/Gemfile.lock`. It is
likely that you need to grant write permissions for that path.

The jekyll/jekyll images had added a new user and group jekyll:jekyll, so we
must copy file with the right permissions.